Mijenjali svasta skupa, nema vremena za vic.
git-svn-id: http://morgoth.zemris.fer.hr/svn/vhdllab/vhdllab/trunk/vhdllab@611 cbae1e92-611c-0410-ab7b-b19ac7dee622

+++ src/web/onClient/hr/fer/zemris/vhdllab/applets/schema2/misc/Time.java
@@ -126,7 +126,7 @@ public final class Time {
 
 	@Override
 	public final String toString() {
-		return (long) this.timeInterval + " " + this.timeMetric.toString();
+		return (double) this.timeInterval + " " + this.timeMetric.toString();
 	}
 
 }
